fix: modify route for MockAdminRepository
